fix: refactor tests to not use JoinFactory and FYTokenFactory
