+++ test/022_join_flash.ts
@@ -9,12 +9,12 @@ import { Result } from '@ethersproject/abi'
 
 const { WAD } = constants
 
-import JoinFactoryArtifact from '../artifacts/contracts/JoinFactory.sol/JoinFactory.json'
+import JoinFactoryArtifact from '../artifacts/contracts/mocks/JoinFactoryMock.sol/JoinFactoryMock.json'
 import ERC20MockArtifact from '../artifacts/contracts/mocks/ERC20Mock.sol/ERC20Mock.json'
 import FlashBorrowerArtifact from '../artifacts/contracts/mocks/FlashBorrower.sol/FlashBorrower.json'
 
 import { Join } from '../typechain/Join'
-import { JoinFactory } from '../typechain/JoinFactory'
+import { IJoinFactory } from '../typechain/IJoinFactory'
 import { ERC20Mock } from '../typechain/ERC20Mock'
 import { FlashBorrower } from '../typechain/FlashBorrower'
 
@@ -30,7 +30,7 @@ describe('Join - flash', function () {
   let otherAcc: SignerWithAddress
   let other: string
   let join: Join
-  let joinFactory: JoinFactory
+  let joinFactory: IJoinFactory
   let token: ERC20Mock
   let borrower: FlashBorrower
 
@@ -52,7 +52,7 @@ describe('Join - flash', function () {
 
   beforeEach(async () => {
     token = (await deployContract(ownerAcc, ERC20MockArtifact, ['MTK', 'Mock Token'])) as ERC20Mock
-    joinFactory = (await deployContract(ownerAcc, JoinFactoryArtifact, [])) as JoinFactory
+    joinFactory = (await deployContract(ownerAcc, JoinFactoryArtifact, [])) as IJoinFactory
     await joinFactory.grantRoles([id(joinFactory.interface, 'createJoin(address)')], owner)
 
     join = (await ethers.getContractAt(
